requirements.txt
fastapi>=0.110
uvicorn>=0.29
python-nmap>=0.7
python-dotenv>=1.0
